+++ app/Models/Enums/AbilityEnum.php
@@ -14,6 +14,8 @@ enum AbilityEnum: string
     case CocktailsImport = 'cocktails.import';
     case IngredientsRead = 'ingredients.read';
     case IngredientsWrite = 'ingredients.write';
+    case InventoryRead = 'inventory.read';
+    case InventoryWrite = 'inventory.write';
     case BarsRead = 'bars.read';
     case BarsWrite = 'bars.write';
     case MenuRead = 'menu.read';