+++ tests/Feature/Http/PATControllerTest.php
@@ -43,7 +43,12 @@ class PATControllerTest extends TestCase
 
         $response = $this->postJson('/api/tokens', [
             'name' => 'My new token',
-            'abilities' => [AbilityEnum::CocktailsRead->value, AbilityEnum::IngredientsWrite->value],
+            'abilities' => [
+                AbilityEnum::CocktailsRead->value,
+                AbilityEnum::IngredientsWrite->value,
+                AbilityEnum::InventoryRead->value,
+                AbilityEnum::InventoryWrite->value,
+            ],
             'expires_at' => Carbon::now()->addMonth()->toAtomString(),
         ]);
 
@@ -51,7 +56,15 @@ class PATControllerTest extends TestCase
         $response->assertJson(
             fn (AssertableJson $json) => $json->has('data.token')
         );
-        $this->assertDatabaseHas('personal_access_tokens', ['name' => 'My new token', 'abilities' => json_encode([AbilityEnum::CocktailsRead->value, AbilityEnum::IngredientsWrite->value])]);
+        $this->assertDatabaseHas('personal_access_tokens', [
+            'name' => 'My new token',
+            'abilities' => json_encode([
+                AbilityEnum::CocktailsRead->value,
+                AbilityEnum::IngredientsWrite->value,
+                AbilityEnum::InventoryRead->value,
+                AbilityEnum::InventoryWrite->value,
+            ]),
+        ]);
     }
 
     public function test_create_token_with_menu_abilities(): void